Support checking for usability of files before confirming selection in picker
If the backend sets the `check` API call details for a file, the frontend will
call it after the user clicks "Select" to check if that file is usable and
disable any errors found. This can be used to warn of eg. a video that is
missing a transcript or a PDF that does not contain selectable text.